requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
blis==0.7.11
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.3
contourpy==1.1.1
cycler==0.11.0
cymem==2.0.8
dataclasses-json==0.6.0
dlib==19.24.2
et-xmlfile==1.1.0
filelock==3.13.1
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
ginza==5.1.3
ginza-transformers==0.4.2
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
googlemaps==4.10.0
greenlet==2.0.2
httplib2==0.22.0
huggingface-hub==0.17.3
idna==3.4
imutils==0.5.4
ja-ginza-electra==5.1.3
Janome==0.5.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
keyboard==0.13.5
kiwisolver==1.4.5
langchain==0.0.300
langcodes==3.3.0
langsmith==0.0.40
logzero==1.7.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mecab-python3==1.0.8
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.6
numpy==1.25.2
openai==0.27.8
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.1
pandas==2.1.3
pathy==0.10.3
pefile==2023.2.7
Pillow==10.0.1
pke @ git+https://github.com/boudinfl/pke.git@69871ffdb720b83df23684fea53ec8776fd87e63
plac==1.4.1
preshed==3.0.9
protobuf==4.21.12
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
pyparsing==3.1.1
PySimpleGUI==4.60.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rake-ja==0.0.1
rake-nltk==1.0.6
regex==2023.10.3
requests==2.31.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
simpleaudio==1.0.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
spacy==3.6.1
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.1.9
SpeechRecognition==3.10.0
SQLAlchemy==2.0.21
srsly==2.4.8
srt==3.5.3
SudachiDict-core==20230927
SudachiPy==0.6.7
SudachiTra==0.1.8
sympy==1.12
tenacity==8.2.3
thinc==8.1.12
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.13.3
torch==2.1.1
torchvision==0.16.1
tqdm==4.65.0
transformers==4.25.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.217
Unidecode==1.3.7
unidic-lite==1.0.8
uritemplate==4.1.1
urllib3==2.0.4
vosk==0.3.45
wasabi==1.1.2
weasel==0.3.4
websockets==11.0.3
yarl==1.9.2
